#74 [refactor] 파일 위치 변경 및 이름 변경
