+++ src/main/java/org/crops/fitserver/domain/mail/constants/MailConstants.java
@@ -1,0 +1,6 @@
+package org.crops.fitserver.domain.mail.constants;
+
+public class MailConstants {
+  public static String ADMIN_MAIL = "crops.project.fit@gmail.com";
+
+}
