+++ src/main/java/org/crops/fitserver/domain/mail/service/MailService.java
@@ -4,5 +4,5 @@ import org.crops.fitserver.domain.mail.dto.MailRequiredInfo;
 import org.crops.fitserver.domain.mail.dto.MailType;
 
 public interface MailService {
-  public void send(MailType mailType, String email, MailRequiredInfo mailRequiredInfo);
+  void send(MailType mailType, String email, MailRequiredInfo mailRequiredInfo);
 }
