+++ src/main/java/org/crops/fitserver/domain/mail/service/MailServiceImpl.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.crops.fitserver.domain.mail.dto.MailRequiredInfo;
 import org.crops.fitserver.domain.mail.dto.MailType;
-import org.crops.fitserver.global.mail.service.MailManager;
+import org.crops.fitserver.global.mail.MailManager;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
@@ -22,7 +22,7 @@ public class MailServiceImpl implements MailService {
       mailManager.send(email, mailType.getSubject(), mailType.replaceBy(mailRequiredInfo));
     } catch (Exception e){
       log.error("Failed to send email", e);
-      //메일 전송 에러가 났으면 무시.
+      //메일 전송 에러가 났으면 무시하고 진행.
     }
 
   }
