+++ src/main/java/org/crops/fitserver/domain/project/service/ProjectServiceImpl.java
@@ -1,5 +1,7 @@
 package org.crops.fitserver.domain.project.service;
 
+import static org.crops.fitserver.domain.mail.constants.MailConstants.ADMIN_MAIL;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.crops.fitserver.domain.mail.dto.AdminMailType;
@@ -79,7 +81,7 @@ public class ProjectServiceImpl implements ProjectService {
         ProjectReportHistory.create(projectMember.getId(), targetProjectMember.getId(), projectId,
             request.reportType(), request.description()));
 
-    mailService.send(AdminMailType.REPORT, targetProjectMember.getUser().getEmail(),
+    mailService.send(AdminMailType.REPORT, ADMIN_MAIL,
         ReportMailRequiredInfo.of(projectMember.getUser(), targetProjectMember.getUser(),
             request.reportType(), request.description()));
   }
